🔨 refactor: suspense 적용 및 error boundy fallback에 workspace 에러 페이지 설정
