+++ apps/client/src/app/App.tsx
@@ -1,9 +1,10 @@
+import { ErrorPage, WorkspaceErrorPage } from '@/pages';
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
-import { ToasterWithMax } from '@/shared/ui';
-import { ErrorPage } from '@/pages/ErrorPage/ErrorPage';
-import { lazy, Suspense } from 'react';
-import { Loading } from '@/shared/ui';
+import { Suspense, lazy } from 'react';
+
 import { Helmet } from 'react-helmet-async';
+import { Loading } from '@/shared/ui';
+import { ToasterWithMax } from '@/shared/ui';
 
 // lazy 로딩
 const HomePage = lazy(() =>
@@ -11,7 +12,7 @@ const HomePage = lazy(() =>
 );
 
 const WorkspacePage = lazy(() =>
-  import('@/pages/Workspacepage/WorkspacePage').then((module) => ({
+  import('@/pages/WorkspacePage/WorkspacePage').then((module) => ({
     default: module.WorkspacePage,
   }))
 );
@@ -47,12 +48,13 @@ const router = createBrowserRouter([
           <title>BooLock - 작업 공간</title>
           <meta name="description" content={`워크스페이스에서 HTML과 CSS를 연습해보세요.`} />
         </Helmet>
+
         <Suspense fallback={<Loading />}>
           <WorkspacePage />
         </Suspense>
       </>
     ),
-    errorElement: <ErrorPage />,
+    errorElement: <WorkspaceErrorPage />,
   },
   {
     path: '*',
