+++ apps/client/src/pages/index.ts
@@ -1,3 +1,5 @@
 export { HomePage } from './HomePage/HomePage';
 export { NotFound } from './NotFound/NotFound';
-export { WorkspacePage } from './Workspacepage/WorkspacePage';
+export { WorkspacePage } from './WorkspacePage/WorkspacePage';
+export { ErrorPage } from './ErrorPage/ErrorPage';
+export { WorkspaceErrorPage } from './WorkspaceErrorPage/WorkspaceErrorPage';
